Initial commit
Scraping of files appears to work.  Files are currently added to a
sqlite database, which may be changed to a MySQL soon for performance
reasons.  At the moment, no pdftotext processing is being done, and
the main loop file (test.py) is, as the name implies, only intended
for testing.  Still got a bunch of TODOs in there, too, but it
basically works right now.