Return path on existing file
If the file to be downloaded already exists, the code will now return
the path to the file (instead of None).  This allows us to more
easily add files that have been downloaded, but not added to the
database - for example after a crash.  Files that are already in
the database will not be processed a second time, as the metadata
code will ignore any file that's already in the database.